fix(engine,jit,stripper,parser,runtime,modules,sdk,docs): 8th fresh-eyes pass — 88 confirmed, 40+ fixed
20-dimension workflow scan + adversarial verify surfaced 88 confirmed
defects across the full codebase; 40+ fixed in this pass.

Critical fixes:
- zigts/build.zig: SQLITE_THREADSAFE=0 -> 2 — the HTTP server uses a
  worker-thread pool; zero mutexing corrupted shared SQLite globals
  across concurrent requests
- jit/optimized: lt/lte/gt/gte, lt_const_i8/le_const_i8, if_true/
  if_false now return UnsupportedOpcode and fall back to baseline, which
  tag-guards operands; the old path sign-extended raw NaN-boxed payload
  bits and compared as i32, producing silently wrong booleans on float,
  string, object, or falsy operands in hot functions
- tools/file_io: apply_edit is now atomic (write temp + fsync + rename);
  a crash mid-write no longer corrupts the user's handler file

High-severity fixes:
- stripper: as/satisfies as plain JS identifiers no longer eat following
  whitespace — save/restore pos/col/line around skipWhitespaceTracked()
  on the false branch (mirroring tryStripColonAnnotation); drifted every
  downstream byte offset by one per occurrence
- stripper: non-null assertion (!), multiline union/intersection type
  aliases, and object/array default param values now stripped correctly
- tokenizer: TokenizerState / saveState / restoreState now include
  subst_brace_depths[16]u8 — a single-token lookahead crossing a { or }
  inside a template substitution no longer permanently corrupts depth
- tokenizer: escaped newline (\ + newline) in string/template literals
  now increments self.line, keeping later source locations accurate
- parser: destructured arrow params ({a,b} / [a,b] in param position)
  emit a clear unsupported_feature diagnostic instead of "unexpected token"
- validate: fail-closed on unknown constraint keywords (pattern,
  additionalProperties, etc.); typeless schemas with required/properties
  now reject non-object input
- decode: multipart boundary anchored to CRLF + --<boundary> per RFC
  2046; bare substring match could split on embedded boundary in values
- decode: duplicate form field names now rejected (DuplicateField)
- jit/arm64: bcond now validates i19 branch-offset range before patching
- pool: release() uses cmpxchgStrong — spurious CAS no longer destroys
  a reusable runtime
- object: prototype chain walk has depth cap (1000)
- arena: overflow path now satisfies alignment > 8
- interpreter/arith: concatNValues asserts count <= 16
- interpreter: sound-mode fault in not/if_true/if_false propagates via
  error.TypeError instead of being swallowed in nested frames
- module_binding: fillRandom memsets buffer to 0 on fill failure

Runtime / server / CLI / UX fixes:
- server: WebSocket upgrade now unregisters from pool if 101 write fails
- runtime_cli: edge compiled-out check runs before bad-arg reporting;
  edge errors mapped to clean exit(1) messages
- dev_cli: edge dispatch error-to-exit-code mapping added
- runtime_cli: studio compiled-out message uses shared featureCompiledOut
- verify_cli: --trust-key=<hex> (= form) now accepted
- runtime_cli serve --help: --no-env-check, --security-log, --lifecycle added
- cli_help: zigttp auth openai added to help --all Credentials section
- install.sh: missing sha256sum/shasum is now a hard failure

HTTP / modules / SDK fixes:
- url.zig: IPv6 literal authority without port parsed correctly
- http_mod: negotiate() trims OWS before q=, clamps q to [0,1]
- modules/src/**security: stale module-spec source paths updated (all 18)

PI / expert fixes:
- pi_repair_plan: non-global bindings no longer misread as atom indices
- response_assembler (anthropic + openai): SSE block index capped at
  1024 before use as allocation length
- sse_parser (anthropic): empty-data keep-alive records no longer abort
- http_errors: SO_SNDTIMEO set alongside SO_RCVTIMEO
- anthropic/request: empty tool_result content -> "(no output)"

Docs:
- verification.md: remove stale fetchSync/upstream fault row; add
  decodeFormMultipart to Result-producing list
- CLAUDE.md module table: added fetchWithRetry, decodeFormMultipart
- type_checker: arrays >32 elements now infer type from first 32

Deferred (confirmed, not fixed — architectural or risky):
- Optimized-tier deopt reconstruction (sp-before-guard, header-resume)
- Bytecode verifier control-flow insensitivity (needs BB worklist)
- inferType speculative diagnostic emission
- CORS allowlist (API design change)
- ?? mixing with ||/&& SyntaxError
- Live-reload swap atomicity, detached WS thread lifetime

Intent: fix-defect
Scope: engine/jit, engine/stripper, engine/parser, engine/gc,
  runtime/server, runtime/cli, modules/security, modules/http,
  modules/sql, sdk, pi/providers, tools/expert, docs
Decided-Against: parallel per-package fix agents with worktrees
  (merge conflicts on shared headers; single sequential agent safer)
Decided-Against: fixing optimized-tier deopt inline (large
  hand-assembly rework; bail-to-baseline closes the silent-wrong-result
  class with zero risk)
Session: 2026-06-14/fresh-eyes-8th-pass
Context: {"workflow_dimensions":20,"raw_findings":88,"confirmed":75,"fixed":40,"test_suites":"zigts+zruntime+cli green"}

+++ packages/pi/src/ui_payload.zig
@@ -6,7 +6,7 @@ pub const DiagnosticItem = struct {
     severity: []u8,
     path: []u8,
     line: u32,
-    column: u16,
+    column: u32,
     message: []u8,
     introduced_by_patch: ?bool = null,
 
@@ -16,7 +16,7 @@ pub const DiagnosticItem = struct {
         severity: []const u8,
         path: []const u8,
         line: u32,
-        column: u16,
+        column: u32,
         message: []const u8,
         introduced_by_patch: ?bool,
     ) !DiagnosticItem {
@@ -575,7 +575,7 @@ pub const ViolationDeltaItem = struct {
     severity: []u8,
     message: []u8,
     line: u32,
-    column: u16,
+    column: u32,
     introduced_by_patch: bool,
 
     pub fn init(
@@ -585,7 +585,7 @@ pub const ViolationDeltaItem = struct {
         severity: []const u8,
         message: []const u8,
         line: u32,
-        column: u16,
+        column: u32,
         introduced_by_patch: bool,
     ) !ViolationDeltaItem {
         return .{
